openApp extracted to AppsMenuUseCase
