+++ app/src/test/java/com/brightbox/hourglass/usecases/AppsMenuUseCaseTest.kt
@@ -1,7 +1,6 @@
 package com.brightbox.hourglass.usecases
 
 import android.content.Context
-import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
 import io.mockk.MockKAnnotations
 import io.mockk.every
@@ -10,20 +9,20 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Test
 
-class GetInstalledApplicationsUseCaseTest{
+class AppsMenuUseCaseTest{
 
     @RelaxedMockK
     private lateinit var mockContext: Context
 
     private lateinit var mockPackageManager: PackageManager
 
-    private lateinit var getInstalledApplicationsUseCase: GetInstalledApplicationsUseCase
+    private lateinit var appsMenuUseCase: AppsMenuUseCase
 
     @Before
     fun onBefore() {
         MockKAnnotations.init(this) // Inicializa MockK
         every { mockContext.packageManager } returns mockPackageManager // Mockea el PackageManager
-        getInstalledApplicationsUseCase = GetInstalledApplicationsUseCase(mockContext)
+        appsMenuUseCase = AppsMenuUseCase(mockContext)
     }
 
     @Test
